Disable export of revocations_matrix_distribution_by_officer to avoid timeouts (Recidiviz/recidiviz-data#5103)
The metric view export for MO has gotten so long that it sometimes times out after running for a full hour. If the timeout happens when we're copying files from the staging area to the final prod bucket, it can result in some very bad, partially written files.

This is a temporary fix that disables export of the largest file until we a) need it on the frontend and b) have optimized the export such that we can re-enable export without timing out.

GitOrigin-RevId: b026994f2f8f77c30615e39c839a88ca2f7edebd