Run /export/metric_view_data inside of a Cloud Task for better observability (Recidiviz/recidiviz-data#5105)
The `/export/metric_view_data ` is currently called from inside a Cloud Function which times out with a 499 after 10 min. However, the metric view export takes longer than that. The daily 499s are masking real failures that happen after the 10 min mark. This PR moves the execution of the main view export code to a Cloud Task so we have 1 hour of runtime.

In this PR:
* New endpoint `/export/create_metric_view_data_export_task` that merely queues a task to a task queue that will execute `/export/metric_view_data`
* Update the cloud function to call `/export/create_metric_view_data_export_task` instead of `/export/metric_view_data` directly.
* Some refactoring to put shared CloudTaskManager code in the same place

Closes Recidiviz/recidiviz-data#4379

GitOrigin-RevId: 23611e600c26936b1d0f0715b3bfa1d2a2e1ae28