[Direct ingest] Raw table migration collection + execeution (Recidiviz/recidiviz-data#5556)
Builds a mechanism for collecting all `RawDataMigration` objects (implemented in Recidiviz/recidiviz-data#5555)  defined for given region and hooks it into the `DirectIngestRawFileImportManager` so that migrations get run for the appropriate raw data files as soon as they are imported.

**Critically, this PR includes a Copybara exemption for any migrations files in the new `raw_data/migrations` directories, since the migration files themselves may reference person external ids which we cannot expose externally**

Also includes tests that ensure that all region migration files have correct names and reference file tags and columns that are documented.

Work towards Recidiviz/recidiviz-data#4829

GitOrigin-RevId: 7d483f461d3f5e24463c032c8e79049005020d12