[PA ingest] Add migration to remove extra space from Wernersville suboffice name (Recidiviz/recidiviz-data#6165)
The query I used to generate the RECIDIVIZ_REFERENCE_supervision_location_ids improperly handled whitespace for one suboffice, leaving the external_id/name as '07 - Wernersville' (with an extra space before Wernersville).

This PR adds a migration on that raw data reference table to fix up the string to remove whitespace (and also fixes the query in the docs so it won't produce this bad data again).

This will take effect as soon as I manually execute these migrations and we do another fresh metric export.

Closes Recidiviz/pulse-dashboardRecidiviz/recidiviz-data#848

GitOrigin-RevId: 15124b4d8685ec7185907de2eac5ce0be2b38ffd