Dropping single day erroneous incarceration periods in pipelines (Recidiviz/recidiviz-data#6219)
Drops any incarceration periods where the admission_date is the same as the release_date, and where the release_reason is RELEASED_FROM_ERRONEOUS_ADMISSION. This is common in US_PA where people have single-day PAROLE_REVOCATION admissions to an SCI followed by a transfer to a community facility. US_PA is the only state that currently has this kind of period, so this won't impact any other state's output.

GitOrigin-RevId: 1f171522a6358a3697a64893f1ee0e8fe14dd56f