Updating race/gender values (Recidiviz/recidiviz-data#6211)
This makes sure that any gender/race category that is not supported on the FE ends up in the "Other" category. We were already sending the "OTHER" rows for the race chart. We don't currently show the "Other" counts on the FE in the gender/race charts, but it's possible that we will need to in order to make the sums add up to 100% (chatting w Serena about this now).

This fixes all remaining failing Lantern validations. Have tested validations locally for US_PA and US_MO.

GitOrigin-RevId: 54f0dac8b27b32a67ff5b55e3c21d91c2ac648f9