Hack to get supervision site and agent info for US_PA (Recidiviz/recidiviz-data#6316)
Due to an ingest bug and a planned launch, we have to hack the supervision_period_to_agent_association view to send in supervision site information to the supervision pipeline in the agent_external_id field. This PR is the calculation logic for handling this hacked agent_external_id field. We shouldn't merge until we've run a test pipeline with the updated reference view.

Filed Recidiviz/recidiviz-data#6314 to remove this hack once the data is reliable.

GitOrigin-RevId: edbf38b81c8cd26903dddd3237e9d997c326b87e