Add external_id uniqueness constraint to state_charge [skip validation ingest_docs] (Recidiviz/recidiviz-data#6467)
Adds a raw table migration for the last remaining known source of duplicate charge ids in PA. Adds a migration to temporarily add a -dupN suffix to existing duplicates that will be cleaned up with subsequent reruns. Adds migration to introduce a uniqueness constraint to external_id column of the state_charge table.

Closes Recidiviz/recidiviz-data#5480
Closes Recidiviz/recidiviz-data#5490

GitOrigin-RevId: 5686c669d95d3804cd1cd98472c070e17f6a45fa