[Multi-DB] Add database_name col to direct_ingest_ingest_file_metadata (Recidiviz/recidiviz-data#6473)
Adds a `database_name` column to the `direct_ingest_ingest_file_metadata` operations DB table so that ingest will eventually be able to track which ingest view files have been discovered/processed etc in the context of a single database. This will allow us to have primary/secondary databases at different states of ingest for a given state.

This PR threads the database name params up into the  PostgresDirectIngestFileMetadataManager but not past it. There will be follow-up work (tracked in Recidiviz/recidiviz-data#6077) to make ingest pass in the correct DB name based on which bucket's files are being processed.

Closes Recidiviz/recidiviz-data#6076.

GitOrigin-RevId: 8c7185901e54f799c72d8df2ef2a29fa898fb8df