Caching BuildableAttr field types  (Recidiviz/recidiviz-data#6590)
Caches mapping of attributes to the type of field that is stored in the attribute for all classes that use the build_from_dictionary function. Reduced the total running time of a US_PA supervision pipeline from 73.265 vCPU hrs to 63.611 vCPU hrs. The cumulative cpu time spent in the build_from_dictionary function was reduced by 66%.

GitOrigin-RevId: 746a99500172e4554beb230d5098450607667f4e