Calculator step of incarceration and program pipelines return RecidivizMetric objects (Recidiviz/recidiviz-data#6608)
This PR removes the step of storing metric attributes and values in a dictionary and then using the build_from_dictionary to later build the metric classes for the incarceration and program pipelines. Now in these pipelines we return the ...Metric objects instead of tuples with the deprecated "value" attached to the metric combo dict.

GitOrigin-RevId: 6ac9af3b69e8d2a080b7f70d83d6f2349325d0a0