Removing metric dict for recidivism pipeline (Recidiviz/recidiviz-data#6612)
This PR removes the step of storing metric attributes and values in a dictionary and then using the build_from_dictionary to later build the metric classes for the recidivism pipeline. Now in this pipeline we return the ...Metric objects from a metric_producer class instead of tuples with the deprecated "value" attached to the metric combo dict.

GitOrigin-RevId: 66ac6affcb1faa25e7c7af2f7b878ba905545df9