[Ingest] Add helper for deriving `SystemLevel` from `Region` (Recidiviz/recidiviz-data#6697)
Adds functionality to derive the `SystemLevel` directly from the `Region` so we don't have to figure out the ingest controller class first to get the `SystemLevel`. Will simplify some of the work for Recidiviz/recidiviz-data#6077 since the ingest bucket names are generated with `SystemLevel` as input.

Work towards Recidiviz/recidiviz-data#6077.

GitOrigin-RevId: 00c6fd217d157a2e009229bd60e57711ae949df7