[Multi-DB] Add `bucket` param to two ingest endpoints (Recidiviz/recidiviz-data#6736)
This updates the `DirectIngestCloudTaskManager` to send a new `bucket` param to the `/direct/scheduler` and `/direct/handle_new_files` endpoints, which will allow us to identify which `DirectIngestInstance` to run ingest for.

This PR does NOT update the endpoints in `direct_ingest_control.py` to actually parse this param out of the endpoint. For now it will be ignored. Ideally, this PR would land in a prod version before the change that parses the parameter.

Work towards Recidiviz/recidiviz-data#6077

GitOrigin-RevId: cdbf1d7391b4f26161be1bf1ca551f0ca2728940