Remove hard-coded SchemaType.CASE_TRIAGE in a few places (Recidiviz/recidiviz-data#6969)
As I was doing some other Cloud SQL -> BQ export investigation, I got tripped up that a few functions are hard-coded to use the CASE_TRIAGE schema even though they are not really named in any way that would indicate so.

This updates those functions to take schema_type as a param.

GitOrigin-RevId: 20953b03e56f32c664b6b98cceecf4898c08164b