Add `RELEASE_TO_SUPERVISION` release reason (Recidiviz/recidiviz-data#7015)
Adds a StateIncarcerationPeriodReleaseReason value of RELEASE_TO_SUPERVISION for instances when a person is being released from incarceration onto some form of supervision that are distinct from when a person is going through the process of being granted parole. CONDITIONAL_RELEASE will be used only for instances when a person is being released onto parole by the parole board.

GitOrigin-RevId: ee4f79cd4c7855f9b69c9b870f5b30981361d595