Add cabability to query for CloudSQL instance region from secrets (Recidiviz/recidiviz-data#7105)
In order to be able to generate `EXTERNAL_QUERY()` views dynamically, we need to know the region that a given CloudSQL instance (and associated connection) lives in. This PR adds `get_cloudsql_instance_region()` and associated tests so we can query for this info.

Queries will eventually look like this and use this function to fill in the `region` param:
```
    SELECT
        *
    FROM EXTERNAL_QUERY("{project_id}.{region}.state_cloudsql", "{full_query};")
```

Work towards Recidiviz/recidiviz-data#6078

GitOrigin-RevId: eef2c64585e7650a81cbc62c014bd633b144cb76