Fixing deploy (Recidiviz/recidiviz-data#7089)
The trigger for the historical pipelines was not working in production because we were checking for changes since the most recent deployed tag after the current tag was already deployed to GCP. This moves when we are checking for new changes to before the deploy. Tested by running parts of the script on their own.

GitOrigin-RevId: 8484c9f4b9dc2be4109ac323980936aa9f2dc6a4