Update dataset overrides generation to handle variable materialized locations (Recidiviz/recidiviz-data#7091)
Refactors dataset overrides generation functionality to:
a) Include datasets for all the views' materialized table locations, if they vary from the view datasets.
b) Expose a function for generating overrides for an arbitrary set of views, for use deploying views to a sandbox that aren't part of the regularly deployed views set.

Also adds tests for this functionality!

Part of Recidiviz/recidiviz-data#6078.

GitOrigin-RevId: 1723b0503d61e9228cd77d20c54f44cf00db1364