[Direct Ingest] Support for latin-1 encoded double-dagger raw data separators (Recidiviz/recidiviz-data#7158)
The PA automatic data dump will have ISO-8859-1 (latin-1) encoded files that use the double-dagger character (`‡`) as a separator.

The separator char is configured in a yaml file and is encoded as UTF-8 when we read it from the config file. However, we need to pass the actual ISO-8859-1 encoded bytes to pandas that it will need to split lines on. Even though we previously supported ISO-8859-1 encoded files, we never have had to use a separator whose byte encoding changes between UTF-8 and ISO-8859-1.

This PR adds support for separator chars whose byte encoding changes between UTF-8 and ISO-8859-1.

Additionally, pandas does not support using the C engine (faster) when a separator is represented in multiple bytes in the UTF-8 encoding, so we also overwrite the `engine` arg when using a double-dagger separator.

Closes Recidiviz/recidiviz-data#7160

GitOrigin-RevId: e608bb094809ff57278a73fada6024a408c0ddc4