Fix dag walker parsing for violation sessions (Recidiviz/recidiviz-data#7212)
The project_id was missing in the join to revocation_sessions_materialized which made the DAG walker fail to parse revocation_sessions as a dependency for the violation_sessions view. So when I went to load the views to sandbox it failed because `revocation_sessions_materialized` hadn't been created yet.

GitOrigin-RevId: b2fe2ae8f08527bbf409611cc1e9041441d575a2