Implement file stream wrapper to normalize CSVs to a standard format (Recidiviz/recidiviz-data#7492)
Adds a wrapper around a stream to a file that has arbitrary column separators and line terminators and normalizes the contents to a format that will be accepted by pandas, since pandas does not currently support parsing files with both complex separators and custom line terminators.

Work towards Recidiviz/recidiviz-data#5594.

GitOrigin-RevId: e10241c8007e69b36e103386d916d836a36a87d7